team report

+++ app/schemas/report.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class SprintStats(BaseModel):
-    total_story_points: int
-    total_tasks: int
-    deadlines_missed: int
+    total_story_points: float
+    total_tasks: float
+    deadlines_missed: float
     average_completion_time: float
